requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
requests>=2.32.5
langchain-core>=0.3.74
streamlit>=1.30.0
groq>=0.4.0
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
